[PATCH 24/50] KVM: SVM: Add KVM_SEV_SNP_LAUNCH_UPDATE command
The KVM_SEV_SNP_LAUNCH_UPDATE command can be used to insert data into the
guest's memory. The data is encrypted with the cryptographic context
created with the KVM_SEV_SNP_LAUNCH_START.

In addition to the inserting data, it can insert a two special pages
into the guests memory: the secrets page and the CPUID page.

While terminating the guest, reclaim the guest pages added in the RMP
table. If the reclaim fails, then the page is no longer safe to be
released back to the system and leak them.

For more information see the SEV-SNP specification.

In some cases, such as detecting whether a page fault should be handled
as a private fault or not, KVM will need to handle things different
versus with the existing KVM_X86_PROTECTED_VM type.

Add a new KVM_X86_SNP_VM to allow for this, along with a helper to query
the vm_type.

Backport Notes:
==============
This patch combines commits:
1 - "KVM: x86: Add KVM_X86_SNP_VM vm_type"
2 - "KVM: SVM: Add KVM_SEV_SNP_LAUNCH command"

Some parts of 1st commit were moved up to earlier patch to follow progress
and PSP calls to catch any PSP firmware issues earlier in the port. Now all
of that patch is applied

Booting SNP Gmem guest may fail for couple reasons
1 - "kvm_set_user_memory_region: KVM_SET_USER_MEMORY_REGION2 failed, slot=0,
    start=0x0, size=0xa0000, flags=0x4, gmem_fd=17, gmem_offset=0x0: Invalid
    argument (22)"

    x This error occurs because some of the qemu devices use guest ram gmemfd
      which is 2GB aligned with KVM_GUEST_MEMFD_ALLOW_HUGEPAGE set, and later
      in kvm_gmem_bind() huge page alignment checks fail for devices that have
      region start of size non huge page aligned. Later patch relaxes alignment
      GMEM alignment checks.
    x For now to run modify qemu function kvm_create_gmemfd() gmem.flags = 0
      must be set prior to calling KVM_CREATE_GUEST_MEMFD VM ioctl.

2 - "sev_snp_launch_finish: SNP_LAUNCH_FINISH ret=-22 fw_error=0 ''"

    x This error occurs because SNP_LAUNCH_FINISH is not implemented yet

Co-developed-by: Michael Roth <michael.roth@amd.com>
Signed-off-by: Michael Roth <michael.roth@amd.com>
Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>