[PATCH 25/50] KVM: SVM: Add KVM_SEV_SNP_LAUNCH_FINISH command
The KVM_SEV_SNP_LAUNCH_FINISH finalize the cryptographic digest and stores
it as the measurement of the guest at launch.

While finalizing the launch flow, it also issues the LAUNCH_UPDATE command
to encrypt the VMSA pages.

If its an SNP guest, then VMSA was added in the RMP entry as
a guest owned page and also removed from the kernel direct map
so flush it later after it is transitioned back to hypervisor
state and restored in the direct map.

Backport Notes:
===============
Launching SNP Guest here starts execution of the guest Qemu should
return an error:
	- KVM: unknown exit reason xx
and kernel log should print:
	- SEV-ES guest requested termination: 0x0:0x0
There is a lot of patches remaining to setup addtional
SNP support to launch the guest, features like GHCB protocol
support, SMP boot and so on.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off-by: Harald Hoyer <harald@profian.com>
Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>
[mdr: always measure BSP first to get consistent launch measurements]